requirements.txt
attrs==20.3.0
bcrypt==3.2.0
boto3==1.17.3
botocore==1.20.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==1.0.2
Cython==0.29.28
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.74.0
gunicorn==20.0.4
h11==0.11.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
loguru==0.6.0
lxml==4.6.2
numpy==1.22.3
pandas==1.4.1
packaging==20.9
pluggy==0.13.1
premailer==3.7.0
py==1.10.0
pycparser==2.20
pydantic==1.9.0
PyJWT==2.0.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pyupbit==0.2.24
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.17.1
testresources==2.0.1
toml==0.10.2
urllib3==1.26.2
uvicorn==0.17.0
uvloop==0.15.1
yagmail==0.14.245
